Add x11/wayland as cargo/meson features/options
