+++ Cargo.toml
@@ -14,13 +14,18 @@ lto = true
 debug = 0
 strip = "debuginfo"
 
+[features]
+default = ["wayland", "x11"]
+wayland = ["gdk-wayland", "gst-plugin-gtk4/wayland"]
+x11 = ["gdk-x11", "gst-plugin-gtk4/x11egl", "gst-plugin-gtk4/x11glx"]
+
 [dependencies]
 adw = { package = "libadwaita", version = "0.7", features = ["v1_6"] }
 anyhow = "1.0.59"
 futures-channel = "0.3.19"
 futures-util = { version = "0.3", default-features = false }
-gdk-wayland = { package = "gdk4-wayland", version = "0.9" }
-gdk-x11 = { package = "gdk4-x11", version = "0.9" }
+gdk-wayland = { package = "gdk4-wayland", version = "0.9", optional = true }
+gdk-x11 = { package = "gdk4-x11", version = "0.9", optional = true }
 gettext-rs = { version = "0.7.0", features = ["gettext-system"] }
 gsettings-macro = "0.2"
 gst = { package = "gstreamer", version = "0.23", features = ["v1_20"] }
@@ -28,9 +33,6 @@ gst-plugin-gif = "0.13"
 gst-plugin-gtk4 = { version = "0.13", features = [
     "dmabuf",
     "gtk_v4_14",
-    "wayland",
-    "x11egl",
-    "x11glx",
 ] }
 gtk = { package = "gtk4", version = "0.9", features = ["gnome_46"] }
 once_cell = "1.19.0"
